Update FE595_Assignment1_kprazdni.py
Good work! I like how detailed the graph is, makes it very easy to read. Perhaps the only thing that I would alter would be the y-axis labels, and move them to the right/ left because the cosine and tangent graph sort of go through them. Or make the green a tad lighter so you can read it better but honestly this graph is pretty perfect!